Save section without estudio.
Save the sections only one time without creating a estudio.
